renamed Form.jsx to Form_Loc.jsx
renamed Form.jsx to Form_Loc.jsx to add a separate form for credit calc to avoid confusion
